# backend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LumiCreate-Pro API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/assets/file/{scene_id}/{asset_type}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Asset File",
        "operationId": "get_asset_file_api_projects__project_id__assets_file__scene_id___asset_type__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          },
          {
            "name": "asset_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/image-engine/precheck": {
      "get": {
        "tags": [
          "image-engine"
        ],
        "summary": "Precheck Workflow",
        "description": "生成前校验工作流：ComfyUI 是否在线 + 节点 class_type 是否齐 + 资源引用统计。",
        "operationId": "precheck_workflow_api_image_engine_precheck_get",
        "parameters": [
          {
            "name": "workflow_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/image-engine/workflow/{workflow_name}": {
      "get": {
        "tags": [
          "image-engine"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_api_image_engine_workflow__workflow_name__get",
        "parameters": [
          {
            "name": "workflow_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/image-engine/workflow-meta/{workflow_name}": {
      "get": {
        "tags": [
          "image-engine"
        ],
        "summary": "Get Workflow Meta",
        "operationId": "get_workflow_meta_api_image_engine_workflow_meta__workflow_name__get",
        "parameters": [
          {
            "name": "workflow_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Name"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "video",
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "image-engine"
        ],
        "summary": "Put Workflow Meta",
        "operationId": "put_workflow_meta_api_image_engine_workflow_meta__workflow_name__put",
        "parameters": [
          {
            "name": "workflow_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowMetaPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/image-engine/generate-stream": {
      "post": {
        "tags": [
          "image-engine"
        ],
        "summary": "Generate Single Stream",
        "operationId": "generate_single_stream_api_image_engine_generate_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/routers__image_engine__SingleGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/image-engine/generate-batch-stream": {
      "post": {
        "tags": [
          "image-engine"
        ],
        "summary": "Generate Batch Stream",
        "description": "Launch len(frames) * gen_count concurrent generation tasks.\nMultiplex all events into one SSE stream.\n\nEvent types: queued | progress | completed | error | batch_done\nEach carries: scene_id, frame_type, slot_index\ncompleted carries: images = [{filename, data(base64), type}]",
        "operationId": "generate_batch_stream_api_image_engine_generate_batch_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/routers__image_engine__BatchGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowMetaPayload": {
        "properties": {
          "node_map": {
            "additionalProperties": true,
            "type": "object",
            "title": "Node Map"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "video"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "node_map"
        ],
        "title": "WorkflowMetaPayload"
      },
...
      "routers__image_engine__BatchGenerateRequest": {
        "properties": {
          "workflow_name": {
            "type": "string",
            "title": "Workflow Name"
          },
          "gen_count": {
            "type": "integer",
            "title": "Gen Count",
            "default": 3
          },
          "negative_prompt": {
            "type": "string",
            "title": "Negative Prompt",
            "default": ""
          },
          "width": {
            "type": "integer",
            "title": "Width",
            "default": 0
          },
          "height": {
            "type": "integer",
            "title": "Height",
            "default": 0
          },
          "frames": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Frames"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "workflow_name",
          "frames"
        ],
        "title": "BatchGenerateRequest"
      },
...
      "routers__image_engine__SingleGenerateRequest": {
        "properties": {
          "workflow_name": {
            "type": "string",
            "title": "Workflow Name"
          },
          "positive_prompt": {
            "type": "string",
            "title": "Positive Prompt"
          },
          "negative_prompt": {
            "type": "string",
            "title": "Negative Prompt",
            "default": ""
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "width": {
            "type": "integer",
            "title": "Width",
            "default": 0
          },
          "height": {
            "type": "integer",
            "title": "Height",
            "default": 0
          },
          "scene_id": {
            "type": "string",
            "title": "Scene Id",
            "default": ""
          },
          "frame_type": {
            "type": "string",
            "title": "Frame Type",
            "default": ""
          },
          "slot_index": {
            "type": "integer",
            "title": "Slot Index",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "workflow_name",
          "positive_prompt"
        ],
        "title": "SingleGenerateRequest"
      },
...
    }
  }
}